style: 패키지명 변경 game -> money
