+++ src/main/java/controller/BlackjackRunner.java
@@ -5,8 +5,8 @@ import static view.Command.YES;
 import domain.Deck;
 import domain.DeckGenerator;
 import domain.ExceptionHandler;
-import domain.game.Money;
-import domain.game.PlayersMoney;
+import domain.money.Money;
+import domain.money.PlayersMoney;
 import domain.user.Dealer;
 import domain.user.GameUsers;
 import domain.user.Player;
