+++ src/test/java/domain/money/GameResultTest.java
@@ -1,0 +1,18 @@
+package domain.money;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+
+class GameResultTest {
+    @ParameterizedTest()
+    @CsvSource(value = {"1,LOSE", "2,DRAW", "3,WIN"})
+    @DisplayName("두 결과를 비교하여 승패를 판단한다.")
+    void compareTest(int current, GameResult gameResult) {
+        int opponentValue = 2;
+
+        assertThat(GameResult.compare(current, opponentValue)).isEqualTo(gameResult);
+    }
+}
